Upgrade GeoWebCache build to Java 17
* Upgrade build to use Java 17 as the baseline version.
* Configure maven-enforcer-plugin to require Java 17.
* Upgrade GitHub Actions workflows to use Java 17.
* Update documentation for Java 17, including developer and user guides.
* Upgrade Jetty from 9.x to 10.0.25, the last version to support `javax.servlet`.
* Replace `com.google.common.base.Objects.equal()` with `java.util.Objects.equal()`.
* Update test configurations to use `localhost:8080` instead of deprecated demo servers.
* Fix REST integration tests and address other issues found during local testing on Java 17.
* Increase heap size for the QA build profile.
* Fix swapped Assert.equals arguments

Plugin version upgrades:
* maven-jar-plugin 2.4 -> 3.0.2
* maven-enforcer-plugin 3.0.0-M3 -> 3.5.0
* maven-compiler-plugin 3.10.1 -> 3.14.0
* maven-surefire-plugin 2.22.1 -> 3.5.3
* maven-failsafe-plugin 3.3.1 -> 3.5.3
* maven-war-plugin 3.3.2 -> 3.4.0
* jetty-maven-plugin 9.4.14.v20181114 -> 10.0.25
* maven-javadoc-plugin 3.4.1 -> 3.11.2
* maven-checkstyle-plugin 3.1.1 -> 3.6.0

+++ geowebcache/core/src/main/java/org/geowebcache/storage/blobstore/file/DefaultFilePathFilter.java
@@ -114,8 +114,8 @@ public class DefaultFilePathFilter implements FilenameFilter {
         String[] parts = name.split("_");
         long halfX = Long.parseLong(parts[0]);
         long halfY = Long.parseLong(parts[1]);
-        long shift = z / 2;
-        long half = 2 << shift;
+        long shift = (long) z / 2;
+        long half = 2L << shift;
         long minX = halfX * half;
         long minY = halfY * half;
         long maxX = minX + half;
